requirements.txt
# Django
Django==6.0.1
django-stubs
djangorestframework-stubs
djangorestframework==3.16.1
django-cors-headers==4.9.0

# 图像处理
Pillow==12.1.0

# 客户端 SDK 依赖
requests==2.32.3

# 开发工具
python-dotenv==1.0.1
